add main + better code separation
